pytest: Separate task schedule tests
